+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Add extension fields support for Python
+ - Add extension fields support for TypeScript
 
 ## [0.6.0]